Continued pixel classifier refactoring
Moves variable importance logging to RTrees, to avoid duplication across pixel and object classifiers.
