+++ qupath-extension-processing/src/main/java/qupath/process/gui/commands/ml/PixelClassifierUtils.java
@@ -1,0 +1,159 @@
+package qupath.process.gui.commands.ml;
+
+import ij.CompositeImage;
+import java.awt.image.BufferedImage;
+import java.io.IOException;
+import javafx.beans.property.DoubleProperty;
+import javafx.scene.control.Spinner;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import qupath.fx.dialogs.Dialogs;
+import qupath.imagej.gui.IJExtension;
+import qupath.imagej.tools.IJTools;
+import qupath.lib.common.GeneralTools;
+import qupath.lib.gui.viewer.QuPathViewer;
+import qupath.lib.gui.viewer.overlays.PixelClassificationOverlay;
+import qupath.lib.images.ImageData;
+import qupath.lib.images.servers.ImageServer;
+import qupath.lib.images.servers.ImageServerMetadata;
+import qupath.lib.regions.RegionRequest;
+import qupath.lib.roi.RectangleROI;
+import qupath.opencv.ops.ImageOp;
+import qupath.opencv.ops.ImageOps;
+
+class PixelClassifierUtils {
+
+    private static final Logger logger = LoggerFactory.getLogger(PixelClassifierUtils.class);
+
+    /**
+     * Check whether two servers have the same number of channels, with the same names and in the same order.
+     * Note that the color of the channel does <i>not</i> matter.
+     * @param server the first image
+     * @param server2 the second image
+     * @return true if the channels match, false otherwise
+     */
+    static boolean compatibleChannels(ImageServer<?> server, ImageServer<?> server2) {
+        if (server == null || server2 == null || server.nChannels() != server2.nChannels())
+            return false;
+        if (server == server2)
+            return true;
+        for (int c = 0; c < server.nChannels(); c++) {
+            if (!server.getChannel(c).getName().equals(server2.getChannel(c).getName()))
+                return false;
+        }
+        return true;
+    }
+
+    /**
+     * Bind a {@link DoubleProperty} to a spinner bidirectionally.
+     * This is a convenience method because the required process is non-obvious
+     * and slightly convoluted, since an object property needs to be created and
+     * retained to prevent premature garbage collection.
+     * @param value the property to bind
+     * @param spinner the spinner
+     */
+    static void bindBidirectional(DoubleProperty value, Spinner<Double> spinner) {
+        // Hack to prevent object property being garbage collected
+        var objProperty = value.asObject();
+        spinner.getValueFactory().valueProperty().bindBidirectional(objProperty);
+        spinner.getProperties().put(PixelClassifierPane.class.getSimpleName() + "_BOUND", objProperty);
+    }
+
+    /**
+     * Show features as an ImageJ stack, corresponding to what is visible at the center of the current viewer.
+     * @param viewer the viewer; if this does not contain an image, then nothing will be shown
+     * @param helper the training data helper, used to determine features and resolution
+     * @param preprocessingOp optional preprocessing op to apply to the features (may be null)
+     */
+    static void showImageJFeatureStack(QuPathViewer viewer, PixelClassifierTraining helper, ImageOp preprocessingOp) {
+        ImageData<BufferedImage> imageData = viewer.getImageData();
+        if (imageData == null)
+            return;
+        double cx = viewer.getCenterPixelX();
+        double cy = viewer.getCenterPixelY();
+
+        var op = helper.getFeatureOp();
+        if (preprocessingOp != null)
+            op = op.appendOps(preprocessingOp);
+
+        try (var featureServer = ImageOps.buildServer(imageData, op, helper.getResolution())) {
+            double downsample = featureServer.getDownsampleForResolution(0);
+            int tw = (int)(featureServer.getMetadata().getPreferredTileWidth() * downsample);
+            int th = (int)(featureServer.getMetadata().getPreferredTileHeight() * downsample);
+            int x = (int) GeneralTools.clipValue(cx - tw/2.0, 0, featureServer.getWidth() - tw);
+            int y = (int)GeneralTools.clipValue(cy - th/2.0, 0, featureServer.getHeight() - th);
+            var request = RegionRequest.createInstance(
+                    featureServer.getPath(),
+                    downsample,
+                    x, y, tw, th, viewer.getZPosition(), viewer.getTPosition());
+
+            var imp = IJTools.convertToImagePlus(featureServer, request).getImage();
+
+            CompositeImage impComp = new CompositeImage(imp, CompositeImage.GRAYSCALE);
+            impComp.setDimensions(imp.getStackSize(), 1, 1);
+            for (int s = 1; s <= imp.getStackSize(); s++) {
+                impComp.setPosition(s);
+                impComp.resetDisplayRange();
+            }
+            impComp.setPosition(1);
+            IJExtension.getImageJInstance();
+            impComp.show();
+        } catch (Exception e) {
+            logger.error("Error calculating features", e);
+        }
+    }
+
+    /**
+     * Show the output of a pixel classifier as an ImageJ stack.
+     * @param viewer the viewer
+     * @param overlay an overlay that provides the classifier
+     */
+    static void showImageJClassifierOutput(QuPathViewer viewer, PixelClassificationOverlay overlay) {
+        if (overlay == null) {
+            Dialogs.showErrorMessage("Show output", "No pixel classifier has been trained yet!");
+            return;
+        }
+        var imageData = viewer.getImageData();
+        var server = imageData == null ? null : overlay.getPixelClassificationServer(imageData);
+        if (server == null)
+            return;
+        var selected = viewer.getSelectedObject();
+        var roi = selected == null ? null : selected.getROI();
+        double downsample = server.getDownsampleForResolution(0);
+        RegionRequest request;
+        if (roi == null) {
+            request = RegionRequest.createInstance(
+                    server.getPath(), downsample,
+                    0, 0, server.getWidth(), server.getHeight(), viewer.getZPosition(), viewer.getTPosition());
+        } else {
+            request = RegionRequest.createInstance(server.getPath(), downsample, selected.getROI());
+        }
+        long estimatedPixels = (long)Math.ceil(request.getWidth()/request.getDownsample()) * (long)Math.ceil(request.getHeight()/request.getDownsample());
+        double estimatedMB = (estimatedPixels * server.nChannels() * (server.getPixelType().getBytesPerPixel())) / (1024.0 * 1024.0);
+        if (estimatedPixels >= Integer.MAX_VALUE - 16) {
+            Dialogs.showErrorMessage("Extract output", "Requested region is too big! Try selecting a smaller region.");
+            return;
+        } else if (estimatedMB >= 200.0) {
+            if (!Dialogs.showConfirmDialog("Extract output",
+                    String.format("Extracting this region will require approximately %.1f MB - are you sure you want to try this?", estimatedMB)))
+                return;
+        }
+
+        try {
+            var pathImage = IJTools.convertToImagePlus(
+                    server,
+                    request);
+            var imp = pathImage.getImage();
+            if (imp instanceof CompositeImage && server.getMetadata().getChannelType() != ImageServerMetadata.ChannelType.CLASSIFICATION) {
+                imp.setDisplayMode(CompositeImage.GRAYSCALE);
+            }
+            if (roi != null && !(roi instanceof RectangleROI)) {
+                imp.setRoi(IJTools.convertToIJRoi(roi, pathImage));
+            }
+            IJExtension.getImageJInstance();
+            imp.show();
+        } catch (IOException e) {
+            logger.error("Error showing output", e);
+        }
+    }
+}
